% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph20.R
\docType{data}
\name{graph20}
\alias{graph20}
\title{20 adjacency matrices from Erdős–Rényi models}
\format{
A \code{list} of 20 adjacency matrices of size \eqn{(28\times 28)}.
}
\usage{
data(graph20)
}
\description{
Simulated list of 20 adjacency matrices of 28 nodes. First 10 are from Erdős–Rényi model with \eqn{p=0.9}, and
the latter 10 are generated using \eqn{p=0.5}. Each element in the list is of size \eqn{(28\times 28)}, symmetric,
having values in \eqn{0} or \eqn{1}, and every diagonal element is set as \eqn{0} in accordance with no self-loop assumption.
}
\details{
Below is the code used to generate \emph{graph20}:
\preformatted{
require(stats)
graph20 = list()
for (i in 1:10){ # type-1 adjacency matrices
  rbin   = rbinom(784,1,0.9)
  mat    = matrix(rbin, nrow=28)
  matout = mat*t(mat)
  diag(matout) = 0
  graph20[[i]]=matout
}
for (i in 11:20){ # type-2 adjacency matrices
  rbin   = rbinom(784,1,0.5)
  mat    = matrix(rbin, nrow=28)
  matout = mat*t(mat)
  diag(matout) = 0
  graph20[[i]]=matout
}
}
}
\keyword{datasets}
